+++ NEWS.md
@@ -1,6 +1,6 @@
 # odbc (development version)
 
-
+* Spark SQL: Correctly enumerate schemas away from the current catalog (@detule, #614)
 * Modify `odbcDataType.Snowflake` to better reflect Snowflake Data Types documentation (@meztez, #599).
 * SQL Server: Specialize syntax in sqlCreateTable to avoid failures when
   writing to (new) local temp tables. (@detule, #601)